+++ server/src/api/paymentsTypes/paymentsTypes.controller.js
@@ -1,34 +1,42 @@
 const { Sequelize } = require("sequelize");
 const db = require("../../db/models/index.js");
+const ErrorHandler = require("../../utils/errorHandler.js");
+const logger = require("../../utils/logger.js");
+const { PaymentsTypesErrors } = require("./paymentsTypes.errors.js");
 
 const PaymentsTypes = db.paymentsTypes;
 
-const getPaymentsTypes = async (req, res) => {
+const getPaymentsTypes = async (req, res, next) => {
   try {
     const paymentsTypes = await PaymentsTypes.findAll();
+
     return res.status(200).json(paymentsTypes);
-  } catch (error) {
-    return res.status(500).send({ message: error.message });
+  } catch (err) {
+    const error = new ErrorHandler(err.message, err.statusCode);
+    logger.error(err);
+    next(error);
   }
 };
 
-const getPaymentTypeById = async (req, res) => {
+const getPaymentTypeById = async (req, res, next) => {
   try {
     const { id } = req.params;
 
     const existingPaymentType = await PaymentsTypes.findByPk(id);
     if (!existingPaymentType) {
-      return res.status(409).json({
-        message: "Conflict: This payment type doesn't exist",
+      return res.status(404).json({
+        message: PaymentsTypesErrors.PAYMENT_TYPE_NOT_FOUND,
       });
     }
 
     return res.status(200).json(existingPaymentType);
-  } catch (error) {
-    return res.status(500).send({ message: error.message });
+  } catch (err) {
+    const error = new ErrorHandler(err.message, err.statusCode);
+    logger.error(err);
+    next(error);
   }
 };
-const createPaymentType = async (req, res) => {
+const createPaymentType = async (req, res, next) => {
   try {
     const name = req.body.name.trim();
 
@@ -42,47 +50,61 @@ const createPaymentType = async (req, res) => {
 
     if (existingPaymentType !== null) {
       return res.status(409).json({
-        message: "Unauthorized: This payment type already exists",
+        message: PaymentsTypesErrors.PAYMENT_TYPE_ALREADY_EXISTS,
       });
     }
 
     const newPaymentType = await PaymentsTypes.create(req.body);
 
     return res.status(201).json(newPaymentType);
-  } catch (error) {
-    return res.status(500).send({ message: error.message });
+  } catch (err) {
+    const error = new ErrorHandler(err.message, err.statusCode);
+    logger.error(err);
+    next(error);
   }
 };
 
-const updatePaymentType = async (req, res) => {
+const updatePaymentType = async (req, res, next) => {
   try {
     const { id } = req.params;
 
     const existingPaymentType = await PaymentsTypes.findByPk(id);
     if (!existingPaymentType) {
-      return res.status(409).json({
-        message: "Conflict: This payment type doesn't exist",
+      return res.status(404).json({
+        message: PaymentsTypesErrors.PAYMENT_TYPE_NOT_FOUND,
       });
     }
     const updatePaymentType = await existingPaymentType.update(req.body);
 
     return res.status(200).json(updatePaymentType);
-  } catch (error) {
-    return res.status(500).send({ message: error.message });
+  } catch (err) {
+    const error = new ErrorHandler(err.message, err.statusCode);
+    logger.error(err);
+    next(error);
   }
 };
 
-const deletePaymentType = async (req, res) => {
+const deletePaymentType = async (req, res, next) => {
   try {
     const { id } = req.params;
+
+    const existingPaymentType = await PaymentsTypes.findByPk(id);
+    if (!existingPaymentType) {
+      return res.status(404).json({
+        message: PaymentsTypes.PAYMENT_TYPE_NOT_FOUND,
+      });
+    }
+
     await PaymentsTypes.destroy({
       where: {
         id,
       },
     });
     res.sendStatus(204);
-  } catch (error) {
-    res.status(500).send({ message: error.message });
+  } catch (err) {
+    const error = new ErrorHandler(err.message, err.statusCode);
+    logger.error(err);
+    next(error);
   }
 };
